pyproject.toml
[project]
name = "dungeonmind"
version = "0.2.0"
description = "Foundation API server for DungeonMind projects"
authors = [
    {name = "Alan Meigs", email = "alan.meigs@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.4",
    "fal-client>=0.5.6",
    "itsdangerous>=2.2.0",
    "authlib>=1.3.2",
    "python-dotenv>=1.0.1",
    "openai>=1.53.0",
    "uvicorn>=0.32.0",
    "google-cloud-firestore>=2.19.0",
    "firebase-admin>=6.6.0",
    "python-multipart>=0.0.17",
    "boto3>=1.35.58",
    "pdf2image>=1.17.0",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "sentence-transformers>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-watch>=4.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "routers",
    "firestore",
    "cloudflare",
    "ruleslawyer",
    "cloudflareR2",
    "cardgenerator",
    "storegenerator"
]


